+++ esutil/bulk_indexer_internal_test.go
@@ -38,8 +38,8 @@ import (
 	"testing"
 	"time"
 
+	estransport "github.com/elastic/elastic-transport-go"
 	"github.com/elastic/go-elasticsearch/v8"
-	"github.com/elastic/go-elasticsearch/v8/estransport"
 )
 
 var defaultRoundTripFunc = func(*http.Request) (*http.Response, error) {
@@ -723,7 +723,7 @@ func TestBulkIndexer(t *testing.T) {
 					disableMetaHeader: false,
 					header: http.Header{
 						"X-Test-User":                []string{"UserValue"},
-						estransport.HeaderClientMeta: []string{"h=shouldntbechanged"},
+						elasticsearch.HeaderClientMeta: []string{"h=shouldntbechanged"},
 					},
 				},
 			},
@@ -737,7 +737,7 @@ func TestBulkIndexer(t *testing.T) {
 					Header:            tt.args.header,
 					Transport: &mockTransport{
 						RoundTripFunc: func(request *http.Request) (*http.Response, error) {
-							headerMeta := request.Header.Get(estransport.HeaderClientMeta)
+							headerMeta := request.Header.Get(elasticsearch.HeaderClientMeta)
 
 							if !reValidation.MatchString(headerMeta) {
 								t.Errorf("Meta Header presence is invalid, got : %s, want : %s", headerMeta, tt.want)
