+++ go.mod
@@ -1,3 +1,10 @@
 module github.com/elastic/go-elasticsearch/v8
 
-go 1.11
+go 1.13
+
+replace github.com/elastic/elastic-transport-go => ../elastic-transport-go
+
+require (
+		github.com/elastic/elastic-transport-go v0.0.0
+)
+
